+++ src/components/Learn/LearnSearchBar.tsx
@@ -2,6 +2,7 @@ import { useState } from "react";
 
 import { Icon } from "@/components/ui/icon";
 import { Input, InputGroup } from "@/components/ui/input";
+import { Text } from "@/components/ui/typography";
 import { useAnalytics } from "@/providers/AnalyticsProvider";
 import { TANGLE_WEBSITE_URL } from "@/utils/constants";
 
@@ -19,24 +20,32 @@ export function LearnSearchBar() {
 
   return (
     <InputGroup
-      className="h-11"
+      className="h-11 w-80"
       prefixElement={
         <Icon
           name="Search"
-          size="md"
+          size="lg"
           className="ml-3 text-muted-foreground"
           aria-hidden="true"
         />
       }
+      suffixElement={
+        <div className="flex items-center gap-1 pr-3 text-muted-foreground shrink-0">
+          <Text size="xs" tone="subdued">
+            tangleml.com
+          </Text>
+          <Icon name="ExternalLink" size="xs" aria-hidden="true" />
+        </div>
+      }
     >
       <Input
         variant="noBorder"
-        placeholder="Search docs, tips and examples…"
+        placeholder="Search the docs…"
         value={value}
         onChange={(e) => setValue(e.target.value)}
         onEnter={handleSubmit}
-        aria-label="Search the Learning Hub"
-        className="h-11 text-sm w-80"
+        aria-label="Search the Tangle docs (opens in a new tab)"
+        className="h-11 text-sm"
       />
     </InputGroup>
   );
