+++ src/routes/Dashboard/Learn/LearnHomeView.tsx
@@ -1,6 +1,8 @@
-import { DocumentationPanel } from "@/components/Learn/DocumentationPanel";
+import { DocsQuickLinks } from "@/components/Learn/DocsQuickLinks";
+import { FaqPanel } from "@/components/Learn/FaqPanel";
 import { FeaturedExamples } from "@/components/Learn/FeaturedExamples";
 import { FeaturedTours } from "@/components/Learn/FeaturedTours";
+import { HelpCard } from "@/components/Learn/HelpCard";
 import { LearnPageHeader } from "@/components/Learn/LearnPageHeader";
 import { LearnSearchBar } from "@/components/Learn/LearnSearchBar";
 import { OnboardingHero } from "@/components/Learn/OnboardingHero";
@@ -16,7 +18,10 @@ export function LearnHomeView() {
           description="Everything you need to get the most out of Tangle, all in one place."
           icon="GraduationCap"
         />
-        <LearnSearchBar />
+        <BlockStack gap="3">
+          <LearnSearchBar />
+          <DocsQuickLinks />
+        </BlockStack>
       </BlockStack>
 
       <OnboardingHero />
@@ -28,7 +33,14 @@ export function LearnHomeView() {
 
       <FeaturedExamples />
 
-      <DocumentationPanel />
+      <div className="grid grid-cols-1 lg:grid-cols-4 gap-6 items-start w-full">
+        <div className="lg:col-span-2">
+          <FaqPanel />
+        </div>
+        <div className="lg:col-span-2 w-full flex justify-center lg:sticky lg:top-[calc(50vh-5rem)]">
+          <HelpCard />
+        </div>
+      </div>
     </BlockStack>
   );
 }
